chore: split test folders (#14)
* chore(test): split unit and functional

* chore(test): fix result test
