+++ tests/functional/fixtures/exit-code-without-error-message.sh
