+++ tests/functional/fixtures/return-error-with-code.sh
