+++ tests/functional/fixtures/return-error.sh
