+++ tests/functional/fixtures/simple-with-input.sh
